Zero firing refinements.
Some improvements in zero's firing with weapons:
1. The game removes the inactive objects - dead monsters & projectiles which hit something or which are outside of screen
2. The game process now the projectiles hitting monsters and substract health from the monsters
3. Projectiles are fired at the last Zero's direction he moved & from the Zero's current position.

+++ ResourcesManager.cpp
@@ -52,7 +52,7 @@ void ResourcesManager::addGameObject(GameObject* gameObject)
 
 void ResourcesManager::removeInactiveGameObjects()
 {
-	for (size_t i = m_gameObjects.size() - 1; i >= 0; i--)
+	for (int i = m_gameObjects.size() - 1; i >= 0; i--)
 	{
 		if (!m_gameObjects[i]->isActive())
 			m_gameObjects.erase(m_gameObjects.begin() + i);
